:goal_net: Move Table into react_library
