+++ ark_nova_stats/frontend/src/components/GameLogArchivesTable.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import {getDate} from '../../../../react_library/DateUtils';
 import GameLogArchive from '../types/GameLogArchive';
-import Table from './Table';
+import Table from '../../../../react_library/Table';
 import PageLink from './PageLink';
 
 type GameLogArchivesTableParams = {
