+++ react_library/Table/index.test.tsx
@@ -1,7 +1,7 @@
 import '@testing-library/jest-dom'
 import { render, screen } from '@testing-library/react';
 import React from 'react';
-import Table from './Table';
+import Table from './index';
 
 type TestTableRow = {
     "Some": React.JSX.Element,
